Convert internal error checks from if-raises to asserts
Semantically, asserts are for checking assumptions made by the programmer, which if not fulfilled imply a bug which can be resolved.

error-raising is for cases when the programmer does not know what to do (e.g. a file the user should provide is missing), and the responsibility lies with the user to rectify it. The error tries to be as specific and helpful as possible in helping the user debug, since they would not know the program's inner workings.